changed capSensor addresses to 32-bit unsigned integers instead of 8-bit (they actually are 32-bit

+++ trunk/DiscoDance/ronelibs/roneos/src/Sensors/capSensor.h
@@ -10,7 +10,7 @@
 
 
 void initCapSensor();
-boolean getCapSensor(uint8 base, uint8 pin);
+boolean getCapSensor(uint32 base, uint32 pin);
 
 
 #endif /* CAPSENSOR_H_ */
